Manually merged changes from main
