+++ test/parser/UnknownFiles.test.ts
@@ -1,0 +1,31 @@
+import { getFileMetrics, testFileMetrics } from "./TestHelper";
+import { FileMetric } from "../../src/parser/metrics/Metric";
+import fs from "fs";
+import { GenericParser } from "../../src/parser/GenericParser";
+
+describe("Test for handling files with unknown or no file extension", () => {
+    const unknownTestResourcesPath = "./resources/unknown/";
+
+    describe("Include files with unknown or no file extension", () => {
+        it("should list files with unknown file extension", async () => {
+            const results = await getFileMetrics(unknownTestResourcesPath);
+
+            const filePath = fs.realpathSync(unknownTestResourcesPath + "example.unknownExtension");
+            expect(results.unknownFiles.includes(filePath)).toBe(true);
+        });
+
+        it("should list files with no file extension", async () => {
+            const results = await getFileMetrics(unknownTestResourcesPath);
+
+            const filePath = fs.realpathSync(unknownTestResourcesPath + "ExampleWithoutExtension");
+            expect(results.unknownFiles.includes(filePath)).toBe(true);
+        });
+
+        it("should still list files with known extension", async () => {
+            const results = await getFileMetrics(unknownTestResourcesPath);
+
+            const filePath = fs.realpathSync(unknownTestResourcesPath + "known.java");
+            expect(results.fileMetrics.has(filePath)).toBe(true);
+        });
+    });
+});
